model/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.15.0
numpy==1.26.2
Pillow==10.1.0
opencv-python-headless==4.8.1.78
scikit-learn==1.3.2
matplotlib==3.8.2
